+++ test/wallet/network_info_test.dart
@@ -2,42 +2,13 @@ import 'package:alan/alan.dart';
 import 'package:test/test.dart';
 
 void main() {
-  test('fromJson works properly', () {
-    final json = {'bech32_hrp': 'test', 'full_node_host': 'test.example.com'};
-    final expected = NetworkInfo(
+  test('NetworkInfo toJson and fromJson', () {
+    final info = NetworkInfo.fromSingleHost(
       bech32Hrp: 'test',
-      fullNodeHost: 'test.example.com',
+      host: 'test.example.com',
     );
-    expect(NetworkInfo.fromJson(json), expected);
-  });
-
-  test('toJson works properly', () {
-    final info = NetworkInfo(
-      bech32Hrp: 'test',
-      fullNodeHost: 'test.example.com',
-    );
-    final json = {
-      'bech32_hrp': 'test',
-      'full_node_host': 'test.example.com',
-      'lcd_port': 1317,
-      'grpc_port': 9090,
-    };
-    expect(info.toJson(), json);
-  });
-
-  test('toString works properly', () {
-    final info = NetworkInfo(
-      bech32Hrp: 'test',
-      fullNodeHost: 'test.example.com',
-    );
-    expect(
-      info.toString(),
-      '{ '
-      'bech32Hrp: test, '
-      'fullNodeHost: test.example.com, '
-      'lcdPort: 1317, '
-      'gRPCPort: 9090 '
-      '}',
-    );
+    final json = info.toJson();
+    final fromJson = NetworkInfo.fromJson(json);
+    expect(fromJson, equals(info));
   });
 }