+++ test/wallet/wallet_test.dart
@@ -7,7 +7,7 @@ import 'package:hex/hex.dart';
 import 'package:test/test.dart';
 
 void main() {
-  final networkInfo = NetworkInfo(bech32Hrp: 'cosmos', fullNodeHost: '');
+  final networkInfo = NetworkInfo.fromSingleHost(bech32Hrp: 'cosmos', host: '');
 
   final testVectors = {
     'cosmos1huydeevpz37sd9snkgul6070mstupukw00xkw9':
@@ -47,7 +47,8 @@ void main() {
   });
 
   test('random generates different wallets', () {
-    final info = NetworkInfo(bech32Hrp: 'cosmos', fullNodeHost: 'example.com');
+    final info =
+        NetworkInfo.fromSingleHost(bech32Hrp: 'cosmos', host: 'example.com');
     final wallets = List.generate(20, (index) => Wallet.random(info));
 
     final map = HashMap.fromIterable(
@@ -87,14 +88,14 @@ void main() {
     ];
     final cosmosAddr = 'cosmos19c506umkrd4ptva9r3gjy7afmjnr4mlgalfmu0';
     final cosmosInfo =
-        NetworkInfo(bech32Hrp: 'cosmos', fullNodeHost: 'example.com');
+        NetworkInfo.fromSingleHost(bech32Hrp: 'cosmos', host: 'example.com');
 
     final wallet = Wallet.derive(mnemonic, cosmosInfo);
     expect(wallet.bech32Address, cosmosAddr);
 
     final desmosAddr = 'desmos19c506umkrd4ptva9r3gjy7afmjnr4mlgf8ytth';
     final desmosInfo =
-        NetworkInfo(bech32Hrp: 'desmos', fullNodeHost: 'example.com');
+        NetworkInfo.fromSingleHost(bech32Hrp: 'desmos', host: 'example.com');
 
     final converted = Wallet.convert(wallet, desmosInfo);
     expect(converted.bech32Address, desmosAddr);
@@ -115,7 +116,7 @@ void main() {
   });
 
   test('sign generates deterministic signatures', () {
-    final info = NetworkInfo(bech32Hrp: 'did:com:', fullNodeHost: '');
+    final info = NetworkInfo.fromSingleHost(bech32Hrp: 'did:com:', host: '');
     final mnemonic = [
       'will',
       'hard',
@@ -151,7 +152,7 @@ void main() {
   });
 
   test('derive with different derivation path works properly', () {
-    final info = NetworkInfo(bech32Hrp: 'desmos', fullNodeHost: '');
+    final info = NetworkInfo.fromSingleHost(bech32Hrp: 'desmos', host: '');
     final mnemonic = [
       'roast',
       'stomach',